requirements.txt
Flask
python-dotenv
openai
gspread
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytesseract
Pillow
requests